requirements.txt
ultralytics
supervision
opencv-python
pillow
fastapi
pydantic
loguru